+++ CHANGELOG.md
@@ -17,6 +17,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Skip notification processing when no read receipt is set
+- Restart notification processing for a room when the read receipt event is in a future sync
+
 ### Security
 
 ## 5.5.0